Reorganize Lorre main application operations to reduce the app body
 * create a package object to define construction of multiple ops from
   resources
 * separate the lorre indexing loop to an independent object
 * rename several concepts to improve consistency and understanding
