+++ build.sbt
@@ -64,7 +64,7 @@ libraryDependencies ++= Seq(
   "com.typesafe.slick"           %% "slick"                         % slickVersion exclude ("org.reactivestreams", "reactive-streams") exclude ("com.typesafe", "config") exclude ("org.slf4j", "slf4j-api"),
   "com.typesafe.slick"           %% "slick-hikaricp"                % slickVersion exclude ("org.slf4j", "slf4j-api"),
   "com.typesafe.slick"           %% "slick-codegen"                 % slickVersion,
-  "com.github.tminglei"          %% "slick-pg"                     % "0.18.0",
+  "com.github.tminglei"          %% "slick-pg"                      % "0.18.0",
   "com.kubukoz"                  %% "slick-effect"                  % "0.1.0" exclude ("com.typesafe.slick", "slick"),
   "org.postgresql"               % "postgresql"                     % "42.1.4",
   "com.github.scopt"             %% "scopt"                         % "4.0.0-RC2",
@@ -79,7 +79,7 @@ libraryDependencies ++= Seq(
   "org.scalatest"                %% "scalatest"                     % "3.0.5" % "it, test",
   "com.stephenn"                 %% "scalatest-json-jsonassert"     % "0.0.3" % "it, test",
   "org.scalamock"                %% "scalamock"                     % "4.1.0" % "it, test",
-  "org.testcontainers"           %  "postgresql"                    % "1.12.3" % "it, test",
+  "org.testcontainers"           % "postgresql"                     % "1.12.3" % "it, test",
   "com.softwaremill.diffx"       %% "diffx-scalatest"               % "0.3.15" % "it, test",
   "org.http4s"                   %% "http4s-blaze-client"           % http4sVersion % IntegrationTest,
   "org.http4s"                   %% "http4s-dsl"                    % http4sVersion % IntegrationTest,
