+++ src/test/scala/tech/cryptonomic/conseil/tezos/JsonDecodersTestFixtures.scala
@@ -1042,8 +1042,6 @@ trait OperationsJsonData {
       )
     )
 
-
-
   val originationJson =
     s"""{
     |  "kind": "origination",
